refactor(MoodItemRow.tsx): optimize code structure and add swipe-to-delete functionality feat(MoodItemRow.tsx): implement swipe-to-delete functionality using react-native-gesture-handler and react-native-reanimated libraries The code in MoodItemRow.tsx has been refactored to improve code structure and readability. The swipe-to-delete functionality has been added using the react-native-gesture-handler and react-native-reanimated libraries. Users can now swipe left on a mood item to delete it.
style(MoodItemRow.tsx): decrease font size of text and moodDescription to 14
style(MoodItemRow.tsx): add DeleteText style with color and font size
The font size of the text and moodDescription in the MoodItemRow component has been decreased to 14 for better visual consistency. Additionally, a new style called DeleteText has been added with a color of theme.colorPurple and a font size of 12.