chore(ios): update Podfile.lock with new dependencies The Podfile.lock has been updated to include the following new dependencies: - RNGestureHandler (2.14.1) - RNReanimated (3.6.2)
These dependencies have been added to support specific features in the iOS application.

chore(ios): update Podfile.lock with new versions of RNGestureHandler and RNReanimated
The Podfile.lock has been updated to include new versions of RNGestureHandler and RNReanimated. This ensures that the project is using the latest versions of these dependencies, which may include bug fixes or new features.